Changes to tests to pass on Rails 4.2.0 Two types of tests were broken: One checking that underscores/spaces could be properly saved, and that the column defaults were correct.
In the case of underscores/spaces, they were broken because the
number was too large to fit into a 4-byte integer. Since we were not
testing size, but format, it seems OK to modify the size while
retaining the formatting.

In the case of defaults, Rails 4.2 changed
`ActiveRecord::ConnectionAdapters::Column#default` to return a
string, so it seemed reasonable to convert that return value to an
integer, since we do not know if that is a bug or a feature, and
since what we really care about is if the resulting ActiveRecord
object has the correct value, which it does.