Enabled local gem caching as default proposed configuration
This sets the default to redirect strategy unless setup is executed.
Then it proposed to use caching as the strategy, and then it ensures
that the storage folder exists so it can fetch gems.

I've pushed the building logic to an Strategies object that handles
dispatching the construction of it.

Also the strategies have better building defaults so they can be built
using the configuration instead of requiring the caller to send all the
arguments. This way it all "just works" when building the Web server.