Logging works as expected, all the layers are writing in the same file.
This was tricky, puma is expecting to work with $stdout and $stderr to log
"events", something where it can "write" to.

On the other hand, rack is also expecting to have something to what is
can "puts" to, again $stdout/$stderr.

Every place I looked for this, whoever was speaking was saying that the
way to go is just to $stdout.reopen("file"), this is ok-ish, but this
does not provide any easy way to rotate the log.

After investigating how the stdlib Logger works, I realized that the
object that handles the rotation of the log is Logger::LogDevice, so the
way to go was to just monkey patch Logger to obtains the LogDevice,
which is almost an IO, and monkey patch that other one just to provide
those 2 final methods that it is looking for: puts and sync=

With this, the rotation will still happen inside the same object, so we
are still using the stdlib Logger.