Fix some broken links in readme
The gemstash readme includes a section titled "Deep Dive" that
includes topics for further reading. Two of those topics, "Private
Gems" and "Multiple Gem Sources", have their own readmes. In those
readmes is a reference to the Gemstash readme's Quickstart section; this
link is broken. And, this commit fixes that.